tests/openfaas/cat-dog/requirements.txt
numpy==1.26.4
torch==2.0.1
torchvision==0.15.2
Pillow==10.0.0
